requirements.txt
# sub-dependencies
-r requirements-deps.txt
# explicit packages for project:
cacheout==0.13.1
celery==5.2.2
certifi==2020.12.5
defusedxml==0.7.1
Django==2.2.28
django-auth-ldap==2.4.0
django-axes==5.27.0
django-bootstrap3==15.0.0
django-celery-beat==2.2.1
django-debug-toolbar==3.2.2
django-environ==0.9.0
django-extensions==3.1.3
django-extra-fields==3.0.2
django-filter==2.4.0
django-jsonfield==1.4.1
django-mptt==0.12.0
django-redis==4.12.1
django-stubs-ext==0.5.0
djangorestframework==3.12.4
django-reversion==3.0.9
django-xff==1.3.0
django-xhtml2pdf==0.0.4
dnspython==2.2.1
drf-yasg==1.20.0
exchangelib==3.3.2
gunicorn==20.1.0
ics==0.7
kaleido==0.2.1
openpyxl==3.0.7
paramiko==2.7.2
plotly==4.14.3
premailer==3.8.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-json-logger==2.0.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-mock==1.9.2
sentry-sdk==1.0.0
tnefparse==1.4.0
typing-extensions==4.2.0
whitenoise==5.2.0
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
